separate code in components and fixing css
